+++ CHANGELOG.md
@@ -22,6 +22,7 @@ Documentation for TransferBench is available at
 - Added new "presets" preset to show available presets and their descriptions
 - Added new "envvars" preset to show environment variables that can change TransferBench behavior
 - Adding information on how to run multi-rank with TransferBench, when run with no args
+- Added new "nica2a" preset (NIC all-to-all over GPUs via NIC executors, multi-node)
 
 ### Modified
 - DMA-BUF support enablement in CMake changed to ENABLE_DMA_BUF to be more similar to other compile-time options